adicionando nova feature de votação
