+++ front/lib/complaint/complaint.dart
@@ -1,0 +1,6 @@
+export 'data/repositories/repositories.dart';
+export 'domain/entities/entities.dart';
+export 'domain/repositories/repositories.dart';
+export 'domain/usecases/usecases.dart';
+export 'complaint_page.dart';
+export 'ui/views/views.dart';
