+++ front/lib/complaint/data/repositories/repositories.dart
@@ -1,0 +1,2 @@
+export 'complaint_api_repository.dart';
+export 'vote_api_repository.dart';
