+++ front/lib/complaint/domain/entities/complaint_result.dart
