+++ front/lib/complaint/domain/repositories/complaint_repository.dart
@@ -1,0 +1,5 @@
+import 'package:mc426_front/complaint/complaint.dart';
+
+abstract class ComplaintRepository {
+  Future<ComplaintResult> createComplaint(Complaint complaint);
+}
