+++ front/lib/complaint/domain/repositories/repositories.dart
@@ -1,0 +1,2 @@
+export 'complaint_repository.dart';
+export 'vote_repository.dart';
