+++ front/lib/complaint/domain/repositories/vote_repository.dart
@@ -1,0 +1,3 @@
+abstract class VoteRepository {
+  Future<bool> vote(int complaintId, bool upvote);
+}
