+++ front/lib/complaint/domain/usecases/usecases.dart
@@ -1,0 +1,2 @@
+export 'create_complaint_usecase.dart';
+export 'vote_usecase.dart';
