+++ front/lib/complaint/domain/usecases/vote_usecase.dart
@@ -1,0 +1,11 @@
+import 'package:mc426_front/complaint/domain/repositories/vote_repository.dart';
+
+class VoteUseCase {
+  final VoteRepository repository;
+
+  VoteUseCase(this.repository);
+
+  Future<bool> call(int complaintId, bool upvote) async {
+    return repository.vote(complaintId, upvote);
+  }
+}
