--- front/lib/create_complaint/complaint.dart
@@ -1,6 +1,0 @@
-export 'data/repositories/repositories.dart';
-export 'domain/entities/entities.dart';
-export 'domain/repositories/repositories.dart';
-export 'domain/usecases/usecases.dart';
-export 'complaint_page.dart';
-export 'ui/views/views.dart';
