--- front/lib/create_complaint/data/repositories/repositories.dart
@@ -1,1 +1,0 @@
-export 'complaint_api_repository.dart';
