--- front/lib/create_complaint/domain/repositories/complaint_repository.dart
@@ -1,5 +1,0 @@
-import 'package:mc426_front/create_complaint/complaint.dart';
-
-abstract class ComplaintRepository {
-  Future<ComplaintResult> createComplaint(Complaint complaint);
-}
