--- front/lib/create_complaint/domain/repositories/repositories.dart
@@ -1,1 +1,0 @@
-export 'complaint_repository.dart';
