--- front/lib/create_complaint/domain/usecases/usecases.dart
@@ -1,1 +1,0 @@
-export 'create_complaint_usecase.dart';
