+++ front/lib/injection/injection.dart
@@ -1,9 +1,7 @@
 import 'package:get_it/get_it.dart';
 import 'package:http/http.dart' as http;
 import 'package:mc426_front/authentication/authentication.dart';
-import 'package:mc426_front/create_complaint/data/repositories/complaint_api_repository.dart';
-import 'package:mc426_front/create_complaint/domain/repositories/complaint_repository.dart';
-import 'package:mc426_front/create_complaint/domain/usecases/create_complaint_usecase.dart';
+import 'package:mc426_front/complaint/complaint.dart';
 import 'package:mc426_front/storage/storage_shared.dart';
 
 final getIt = GetIt.instance;
@@ -17,6 +15,8 @@ setupProviders() async {
 
   getIt.registerLazySingleton<ComplaintRepository>(() => ComplaintApiRepository(client));
   getIt.registerLazySingleton<CreateComplaintUsecase>(() => CreateComplaintUsecase(getIt.get<ComplaintRepository>()));
+  getIt.registerLazySingleton<VoteRepository>(() => VoteApiRepository(client));
+  getIt.registerLazySingleton<VoteUseCase>(() => VoteUseCase(getIt.get<VoteRepository>()));
   getIt.registerLazySingleton<AuthenticationRepository>(() => AuthenticationApiRepository(client));
   getIt.registerLazySingleton<SignInUsecase>(() => SignInUsecase(
         getIt.get<AuthenticationRepository>(),
